Rearrange components into more organized folders and subfolders
