+++ imports/routes/components/SkillTreeCommunityView.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
 // Element JSX Component/Layout
-import { SkillTreeCommunityView } from '/imports/ui/components/SkillTreeCommunityView';
-import { CommunityLeaderboardModal } from '/imports/ui/components/CommunityLeaderboardModal.jsx';
+import { SkillTreeCommunityView } from '../../ui/components/SkillTrees/SkillTreeCommunityView';
+import { CommunityLeaderboardModal } from '../../ui/components/SkillTrees/Leaderboard/CommunityLeaderboardModal.jsx';
 
 export const SkillTreeCommunityViewRoutes = [
   {
