+++ imports/ui/App.jsx
@@ -2,8 +2,8 @@ import React, { Suspense } from 'react';
 import { Outlet } from 'react-router-dom';
 
 // JSX UI
-import { NavBar } from '/imports/ui/components/NavBar';
-import { Fallback } from '/imports/ui/components/Fallback';
+import { NavBar } from './components/SiteFrame/NavBar';
+import { Fallback } from './components/SiteFrame/Fallback';
 
 export const App = () => (
   <>
