+++ imports/ui/components/SkillTrees/SkillTreeCommunityView.jsx
@@ -1,8 +1,8 @@
 import React, { useContext } from 'react';
-import { SkillTreeView } from '../components/SkillTreeView';
+import { SkillTreeView } from './SkillTreeView';
 import { SkillTreeCollection } from '/imports/api/collections/SkillTree';
 import { useParams, Outlet, Link, useLocation } from 'react-router-dom';
-import { NavigationDropdown } from '../components/NavigationDropdown';
+import { NavigationDropdown } from './NavigationDropdown';
 import { useFind } from 'meteor/react-meteor-data/suspense';
 import { useSubscribeSuspense } from 'meteor/communitypackages:react-router-ssr';
 import { SubscribeButton } from './SubscribeButton';
