+++ imports/ui/layouts/HelloContainer.jsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { Outlet } from 'react-router-dom';
 
 // JSX UI
-import { Fallback } from '/imports/ui/components/Fallback';
+import { Fallback } from '../components/SiteFrame/Fallback';
 
 export const HelloContainer = () => (
   <>
