+++ imports/ui/layouts/PublicLayout.jsx
@@ -2,7 +2,7 @@ import React, { Suspense } from 'react';
 import { Outlet } from 'react-router-dom';
 
 // JSX UI
-import { Fallback } from '/imports/ui/components/Fallback';
+import { Fallback } from '../components/SiteFrame/Fallback';
 
 export const PublicLayout = () => (
   <>
