+++ imports/ui/pages/GeneralForum.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useRef, useEffect } from 'react';
 import { useParams } from 'react-router-dom';
-import { TopicList } from '/imports/ui/components/TopicListGeneralForum';
+import { TopicList } from '../components/SkillTrees/GeneralForum/TopicListGeneralForum';
 
 export const GeneralForum = () => {
   const { skilltreeId } = useParams();
