+++ imports/ui/pages/SkillTreeCommunity.jsx
@@ -3,7 +3,7 @@ import { Helmet } from 'react-helmet';
 import { Outlet } from 'react-router-dom';
 
 // JSX UI
-import { Fallback } from '/imports/ui/components/Fallback';
+import { Fallback } from '../components/SiteFrame/Fallback';
 
 export const SkillTreeCommunity = () => (
   <>
